Move more parameters utils to metabase-lib (#25900)
* Move subset of parameter-values utils to metabase-lib

* Move parameter mbql utils to metabase-lib

* Create some small target utils

* Move template tag utils to metabase-lib

* Move target/field logic to metabase-lib

* Remove eslint-enable
