+++ frontend/src/metabase-lib/lib/parameters/utils/parameter-values.js
@@ -1,0 +1,20 @@
+export function getValuePopulatedParameters(parameters, parameterValues) {
+  return parameterValues
+    ? parameters.map(parameter => {
+        return parameter.id in parameterValues
+          ? {
+              ...parameter,
+              value: parameterValues[parameter.id],
+            }
+          : parameter;
+      })
+    : parameters;
+}
+
+export function hasDefaultParameterValue(parameter) {
+  return parameter.default != null;
+}
+
+export function hasParameterValue(value) {
+  return value != null;
+}
