+++ frontend/src/metabase-lib/lib/parameters/utils/targets.ts
@@ -1,12 +1,15 @@
 import {
   ParameterTarget,
   ParameterDimensionTarget,
+  ParameterVariableTarget,
 } from "metabase-types/types/Parameter";
+import { SavedCard } from "metabase-types/types/Card";
 import Dimension from "metabase-lib/lib/Dimension";
 import Metadata from "metabase-lib/lib/metadata/Metadata";
 import Question from "metabase-lib/lib/Question";
 import StructuredQuery from "metabase-lib/lib/queries/StructuredQuery";
 import NativeQuery from "metabase-lib/lib/queries/NativeQuery";
+import TemplateTagVariable from "metabase-lib/lib/variables/TemplateTagVariable";
 
 export function isDimensionTarget(
   target: ParameterTarget,
@@ -40,4 +43,30 @@ export function getParameterTargetField(
   }
 
   return null;
+}
+
+export function buildDimensionTarget(dimension: Dimension) {
+  return ["dimension", dimension.mbql()];
+}
+
+export function buildTemplateTagVariableTarget(variable: TemplateTagVariable) {
+  return ["variable", variable.mbql()];
+}
+
+export function buildTextTagTarget(tagName: string) {
+  return ["text-tag", tagName];
+}
+
+export function getTargetFieldFromCard(
+  target: ParameterVariableTarget | ParameterDimensionTarget,
+  card: SavedCard,
+  metadata: Metadata,
+) {
+  if (!card?.dataset_query) {
+    return null;
+  }
+
+  const question = new Question(card, metadata);
+  const field = getParameterTargetField(target, metadata, question);
+  return field ?? null;
 }
