+++ frontend/src/metabase/dashboard/containers/AutomaticDashboardApp.jsx
@@ -9,7 +9,6 @@ import { dissoc } from "icepick";
 import title from "metabase/hoc/Title";
 import withToast from "metabase/hoc/Toast";
 import DashboardData from "metabase/dashboard/hoc/DashboardData";
-import { getValuePopulatedParameters } from "metabase/parameters/utils/parameter-values";
 
 import ActionButton from "metabase/components/ActionButton";
 import Button from "metabase/core/components/Button";
@@ -30,6 +29,7 @@ import Dashboards from "metabase/entities/dashboards";
 import * as Urls from "metabase/lib/urls";
 import * as MetabaseAnalytics from "metabase/lib/analytics";
 import { color } from "metabase/lib/colors";
+import { getValuePopulatedParameters } from "metabase-lib/lib/parameters/utils/parameter-values";
 import * as Q from "metabase-lib/lib/queries/utils/query";
 import Dimension from "metabase-lib/lib/Dimension";
