+++ frontend/src/metabase/lib/pulse.ts
@@ -1,11 +1,7 @@
 import _ from "underscore";
 import MetabaseSettings from "metabase/lib/settings";
 import MetabaseUtils from "metabase/lib/utils";
-import {
-  hasDefaultParameterValue,
-  hasParameterValue,
-  normalizeParameterValue,
-} from "metabase/parameters/utils/parameter-values";
+import { normalizeParameterValue } from "metabase/parameters/utils/parameter-values";
 
 import {
   Channel,
@@ -14,6 +10,10 @@ import {
   Pulse,
   PulseParameter,
 } from "metabase-types/api";
+import {
+  hasDefaultParameterValue,
+  hasParameterValue,
+} from "metabase-lib/lib/parameters/utils/parameter-values";
 
 export const NEW_PULSE_TEMPLATE = {
   name: null,
