+++ frontend/src/metabase/parameters/utils/cards.ts
@@ -1,95 +1,20 @@
 import _ from "underscore";
 
-import {
-  getValuePopulatedParameters,
-  hasParameterValue,
-} from "metabase/parameters/utils/parameter-values";
 import { ParameterWithTarget, UiParameter } from "metabase/parameters/types";
 import { Parameter, ParameterTarget } from "metabase-types/types/Parameter";
 import { Card } from "metabase-types/types/Card";
 import { TemplateTag } from "metabase-types/types/Query";
+import {
+  getValuePopulatedParameters,
+  hasParameterValue,
+} from "metabase-lib/lib/parameters/utils/parameter-values";
 import {
   getParameterTargetField,
   getTemplateTagFromTarget,
 } from "metabase-lib/lib/parameters/utils/targets";
 import Question from "metabase-lib/lib/Question";
 import Metadata from "metabase-lib/lib/metadata/Metadata";
-
-export function getTemplateTagType(tag: TemplateTag) {
-  const { type } = tag;
-  if (type === "date") {
-    return "date/single";
-    // @ts-expect-error -- preserving preexisting incorrect types (for now)
-  } else if (type === "string") {
-    return "string/=";
-  } else if (type === "number") {
-    return "number/=";
-  } else {
-    return "category";
-  }
-}
-
-export function getTemplateTagParameterTarget(
-  tag: TemplateTag,
-): ParameterTarget {
-  return tag.type === "dimension"
-    ? ["dimension", ["template-tag", tag.name]]
-    : ["variable", ["template-tag", tag.name]];
-}
-
-export function getTemplateTagParameter(tag: TemplateTag): ParameterWithTarget {
-  return {
-    id: tag.id,
-    type: tag["widget-type"] || getTemplateTagType(tag),
-    target: getTemplateTagParameterTarget(tag),
-    name: tag["display-name"],
-    slug: tag.name,
-    default: tag.default,
-  };
-}
-
-// NOTE: this should mirror `template-tag-parameters` in src/metabase/api/embed.clj
-
-export function getTemplateTagParameters(
-  tags: TemplateTag[],
-): ParameterWithTarget[] {
-  return tags
-    .filter(
-      tag =>
-        tag.type != null && (tag["widget-type"] || tag.type !== "dimension"),
-    )
-    .map(getTemplateTagParameter);
-}
-
-export function getTemplateTagsForParameters(card: Card) {
-  const templateTags: TemplateTag[] =
-    card &&
-    card.dataset_query &&
-    card.dataset_query.type === "native" &&
-    card.dataset_query.native["template-tags"]
-      ? Object.values(card.dataset_query.native["template-tags"])
-      : [];
-
-  return templateTags.filter(
-    // this should only return template tags that define a parameter of the card
-    tag => tag.type !== "card" && tag.type !== "snippet",
-  );
-}
-
-export function getParametersFromCard(
-  card: Card,
-): Parameter[] | ParameterWithTarget[] {
-  if (!card) {
-    return [];
-  }
-
-  if (card.parameters && !_.isEmpty(card.parameters)) {
-    return card.parameters;
-  }
-
-  const tags = getTemplateTagsForParameters(card);
-  return getTemplateTagParameters(tags);
-}
+import { getParametersFromCard } from "metabase-lib/lib/parameters/utils/template-tags";
 
 export function getCardUiParameters(
   card: Card,
