+++ frontend/src/metabase/parameters/utils/date-formatting.ts
@@ -2,9 +2,9 @@ import { t } from "ttag";
 import _ from "underscore";
 import moment from "moment-timezone";
 
-import { dateParameterValueToMBQL } from "metabase/parameters/utils/mbql";
 import { DATE_OPERATORS } from "metabase/query_builder/components/filters/pickers/DatePicker/DatePicker";
 import { EXCLUDE_OPERATORS } from "metabase/query_builder/components/filters/pickers/DatePicker/ExcludeDatePicker";
+import { dateParameterValueToMBQL } from "metabase-lib/lib/parameters/utils/mbql";
 import { DATE_MBQL_FILTER_MAPPING } from "metabase-lib/lib/parameters/constants";
 import {
   generateTimeFilterValuesDescriptions,
