+++ frontend/src/metabase/parameters/utils/mapping-options.js
@@ -7,13 +7,18 @@ import {
   getTagOperatorFilterForParameter,
   variableFilterForParameter,
 } from "metabase-lib/lib/parameters/utils/filters";
+import {
+  buildDimensionTarget,
+  buildTemplateTagVariableTarget,
+  buildTextTagTarget,
+} from "metabase-lib/lib/parameters/utils/targets";
 
 function buildStructuredQuerySectionOptions(section) {
   return section.items.map(({ dimension }) => ({
     sectionName: section.name,
     name: dimension.displayName(),
     icon: dimension.icon(),
-    target: ["dimension", dimension.mbql()],
+    target: buildDimensionTarget(dimension),
     // these methods don't exist on instances of ExpressionDimension
     isForeign: !!(dimension instanceof ExpressionDimension
       ? false
@@ -26,7 +31,7 @@ function buildNativeQuerySectionOptions(section) {
     name: dimension.displayName(),
     icon: dimension.icon(),
     isForeign: false,
-    target: ["dimension", dimension.mbql()],
+    target: buildDimensionTarget(dimension),
   }));
 }
 
@@ -35,7 +40,7 @@ function buildVariableOption(variable) {
     name: variable.displayName(),
     icon: variable.icon(),
     isForeign: false,
-    target: ["variable", variable.mbql()],
+    target: buildTemplateTagVariableTarget(variable),
   };
 }
 
@@ -44,7 +49,7 @@ function buildTextTagOption(tagName) {
     name: tagName,
     icon: "string",
     isForeign: false,
-    target: ["text-tag", tagName],
+    target: buildTextTagTarget(tagName),
   };
 }
