+++ frontend/src/metabase/parameters/utils/parameter-values.js
@@ -1,25 +1,8 @@
 import { getParameterType } from "metabase-lib/lib/parameters/utils/parameter-type";
-
-export function getValuePopulatedParameters(parameters, parameterValues) {
-  return parameterValues
-    ? parameters.map(parameter => {
-        return parameter.id in parameterValues
-          ? {
-              ...parameter,
-              value: parameterValues[parameter.id],
-            }
-          : parameter;
-      })
-    : parameters;
-}
-
-export function hasDefaultParameterValue(parameter) {
-  return parameter.default != null;
-}
-
-export function hasParameterValue(value) {
-  return value != null;
-}
+import {
+  hasParameterValue,
+  hasDefaultParameterValue,
+} from "metabase-lib/lib/parameters/utils/parameter-values";
 
 export function getParameterValueFromQueryParams(
   parameter,
