+++ frontend/src/metabase/public/components/widgets/AdvancedSettingsPane.jsx
@@ -3,13 +3,13 @@ import React, { useMemo } from "react";
 import { t } from "ttag";
 import cx from "classnames";
 
-import { getValuePopulatedParameters } from "metabase/parameters/utils/parameter-values";
 import Icon from "metabase/components/Icon";
 import Button from "metabase/core/components/Button";
 import ParametersList from "metabase/parameters/components/ParametersList";
 import Select, { Option } from "metabase/core/components/Select";
 
 import { color } from "metabase/lib/colors";
+import { getValuePopulatedParameters } from "metabase-lib/lib/parameters/utils/parameter-values";
 
 import DisplayOptionsPane from "./DisplayOptionsPane";
